+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - Add `autoHighlight` option that will auto highlight the best matching result
 - Feature request: option to hide the drop-down if there are no results found #189
 - feature proposal: keyUp output event #186
+- how to empty search box? #114
 
 ### Fixed bugs
 - Completer doesn't close when the field is cleared #194